Split BriefingPanel and ManagerPanel from ActivityInfoPanel
