+++ src/components/activities/ManagerPanel.tsx
@@ -1,0 +1,38 @@
+import { Paper } from '@respond/components/Material';
+import { OutputForm, OutputText, OutputTime } from '@respond/components/OutputForm';
+import { getActivityStatus } from '@respond/lib/client/store/activities';
+import { Activity, isActive as isParticpantActive, isCheckedIn as isParticpantCheckedIn, Participant, ParticipantStatus } from '@respond/types/activity';
+
+export function ManagerPanel({ activity }: { activity: Activity }) {
+  const reduceActive = (count: number, participant: Participant) => {
+    return count + (isParticpantActive(participant?.timeline[0].status) ? 1 : 0);
+  };
+
+  const reduceStandby = (count: number, participant: Participant) => {
+    return count + (participant?.timeline[0].status === ParticipantStatus.Standby ? 1 : 0);
+  };
+
+  const reduceSignedIn = (count: number, participant: Participant) => {
+    return count + (participant?.timeline[0].status === ParticipantStatus.SignedIn ? 1 : 0);
+  };
+
+  const reduceCheckedIn = (count: number, participant: Participant) => {
+    return count + (isParticpantCheckedIn(participant?.timeline[0].status) ? 1 : 0);
+  };
+
+  return (
+    <Paper elevation={1} sx={{ p: 1 }}>
+      <OutputForm>
+        <OutputText label="State #" value={activity.idNumber} />
+        <OutputText label="Agency" value={activity.organizations[activity.ownerOrgId]?.title} />
+        <OutputText label="Mission Status" value={getActivityStatus(activity)} />
+        <OutputText label="Active Responders" value={Object.values(activity.participants).reduce(reduceActive, 0).toString()}></OutputText>
+        <OutputText label="Standby" value={Object.values(activity.participants).reduce(reduceStandby, 0).toString()}></OutputText>
+        <OutputText label="Responding" value={Object.values(activity.participants).reduce(reduceSignedIn, 0).toString()}></OutputText>
+        <OutputText label="Checked-In" value={Object.values(activity.participants).reduce(reduceCheckedIn, 0).toString()}></OutputText>
+        <OutputTime label="Start Time" time={activity.startTime}></OutputTime>
+        <OutputTime label="End Time" time={activity.endTime}></OutputTime>
+      </OutputForm>
+    </Paper>
+  );
+}
